Update Travis build setup

+++ scripts/travis-build.sh
@@ -4,34 +4,64 @@ set -e
 # Display version of compiler
 $CXX --version
 
-# Store this directory
-DIGRAPHSDIR=$(pwd)
+INITIALDIR=`pwd`
 
 # Download and compile GAP
 cd ..
-git clone -b $GAP_BRANCH --depth=1 https://github.com/$GAP_FORK/gap.git
+git clone -b $GAP_BRANCH --depth=1 https://github.com/$GAP_FORK/gap.git gap
 cd gap
 ./configure --with-gmp=system $GAP_FLAGS
 make
 mkdir pkg
 cd ..
 
-# Compile the Digraphs package
-mv $DIGRAPHSDIR gap/pkg/digraphs
-cd gap/pkg/digraphs
-curl -L -O https://gap-packages.github.io/Digraphs/$DIGRAPHS_LIB.tar.gz
-tar xf $DIGRAPHS_LIB.tar.gz
-rm $DIGRAPHS_LIB.tar.gz
+# Move the package itself to the correct location
+mv $INITIALDIR gap/pkg/$PACKAGE
+cd gap/pkg/$PACKAGE
+
+if [ "$PACKAGE" == "semigroups" ] && [ -d src ]; then
+  cd src
+  git clone -b $LIBSEMIGROUPS_BR --depth=1 https://github.com/james-d-mitchell/libsemigroups.git libsemigroups
+  cd ..
+  ./autogen.sh
+  ./configure $PKG_FLAGS
+  make
+fi
+
+if [ "$PACKAGE" == "digraphs" ]; then
+  curl -L -O https://gap-packages.github.io/Digraphs/$DIGRAPHS_LIB.tar.gz
+  tar xf $DIGRAPHS_LIB.tar.gz
+  rm $DIGRAPHS_LIB.tar.gz
+fi
+
+cd ..
+
+# Download and install dependencies and tools...
+
+# Compile Digraphs
+if [ "$PACKAGE" == "semigroups" ]; then
+  git clone -b $DIGRAPHS_BR --depth=1 https://github.com/gap-packages/Digraphs.git digraphs
+fi
+cd digraphs
 ./autogen.sh
 ./configure $PKG_FLAGS
 make
-cd ../..
+cd ..
 
-# Get the packages
-cd pkg
+# GAPDoc
+echo "Downloading $GAPDOC..."
 curl -O https://www.gap-system.org/pub/gap/gap4/tar.gz/packages/$GAPDOC.tar.gz
 tar xzf $GAPDOC.tar.gz
 rm $GAPDOC.tar.gz
+
+# GenSS
+echo "Downloading $GENSS..."
+curl -O https://www.gap-system.org/pub/gap/gap4/tar.gz/packages/$GENSS.tar.gz
+tar xzf $GENSS.tar.gz
+rm $GENSS.tar.gz
+
+# IO
+echo "Downloading $IO..."
 curl -O https://www.gap-system.org/pub/gap/gap4/tar.gz/packages/$IO.tar.gz
 tar xzf $IO.tar.gz
 rm $IO.tar.gz
@@ -39,6 +69,9 @@ cd $IO
 ./configure $PKG_FLAGS
 make
 cd ..
+
+# Orb
+echo "Downloading $ORB..."
 curl -O https://www.gap-system.org/pub/gap/gap4/tar.gz/packages/$ORB.tar.gz
 tar xzf $ORB.tar.gz
 rm $ORB.tar.gz
@@ -46,8 +79,21 @@ cd $ORB
 ./configure $PKG_FLAGS
 make
 cd ..
+
+# Grape: only install if we're testing the digraphs package
+if [ "$PACKAGE" == "digraphs" ]; then
+  echo "Downloading $GRAPE..."
+  curl -O https://www.gap-system.org/pub/gap/gap4/tar.gz/packages/$GRAPE.tar.gz
+  tar xzf $GRAPE.tar.gz
+  rm $GRAPE.tar.gz
+  cd grape
+  ./configure $PKG_FLAGS
+  make
+  cd ..
+fi
+
 # Only install profiling package if we'll run coverage checks
-if [ ! -z "$COVERAGE" ]; then
+if [ ! -z "$RUNCOVERAGE" ]; then
   echo "Downloading $PROFILING..."
   curl -LO https://github.com/gap-packages/profiling/releases/download/v$PROFILING/profiling-$PROFILING.tar.gz
   tar xzf profiling-$PROFILING.tar.gz
@@ -59,17 +105,10 @@ if [ ! -z "$COVERAGE" ]; then
   make
   cd ..
 fi
-curl -O https://www.gap-system.org/pub/gap/gap4/tar.gz/packages/$GENSS.tar.gz
-tar xzf $GENSS.tar.gz
-rm $GENSS.tar.gz
-curl -O https://www.gap-system.org/pub/gap/gap4/tar.gz/packages/$GRAPE.tar.gz
-tar xzf $GRAPE.tar.gz
-rm $GRAPE.tar.gz
-cd grape
-./configure $PKG_FLAGS
-make
-cd ../../..
 
-# Get gaplint
-echo "Downloading gaplint..."
-git clone -b master --depth=1 https://github.com/james-d-mitchell/gaplint
+# Only install gaplint if we'll run the linting checks
+if [ ! -z "$RUNGAPLINT" ]; then
+  echo "Downloading gaplint..."
+  git clone -b master --depth=1 https://github.com/james-d-mitchell/gaplint gaplint
+  cd ..
+fi
